deactivate include of <avr/cpufunc.h> for toolchain compatibility
Since cpufunc.h is not needed in the context of primary firmware
and since it is not available in all toolchains, the include
of <avr/cpufunc.h> becomes deactivated by github issue-report.
(In case of trouble it remains in sourcecode for reactivation.)

The autor would like to thank Lena-M for reporting this
issue (https://github.com/baerwolf/USBaspLoader/issues/1).

Reported-by: Lena-M
Signed-off-by: Stephan Baerwolf <stephan.baerwolf@tu-ilmenau.de>