fix: "bootloader__do_spm is not located after interrupts" for ATmega2561 and ATmega1281
Since the datasheet is the same as for ATmega1284 and ATmega2560 the depicted
table of interrupt seems to be the same, too.
But on the second impression there are small "*" at some interrupts not
implemented in the affected devices.

So adapt the bootloader__do_spm addresses - thank you oh you dear fail safe.

Without following code added last time, I never would have found it before some
critical fail!

    #if defined(_VECTORS_SIZE)
      #if (funcaddr___bootloader__do_spm != (BOOTLOADER_ADDRESS+_VECTORS_SIZE))
	#error "bootloader__do_spm is not located after interrupts - sth. is very wrong here!"
      #endif
    #endif

Signed-off-by: Stephan Baerwolf <stephan.baerwolf@tu-ilmenau.de>