introduce new feature for the updater: UPDATECRC32
In the head of Makefile.inc you now can configure "UPDATECRC32"
to enable the updater to check integrity of the new firmware image.
The CRC32 of polynomial 0xEDB88320 is used. This CRC is the same,
the non-volatile-memory controller of ATxmega series is using.

Enabling this feature will only increase the update by a few
bytes (~180bytes) , but will offer more safety against bogus images.

By default the CRC is NOT used, but it can be enabled to a
fix value or calculated automatically during compile.

Signed-off-by: Stephan Baerwolf <stephan.baerwolf@tu-ilmenau.de>