fix annoying compiler warning "'usbFunctionDescriptor' used but never defined"
Because in order to save a few bytes of code, the usbdrv.c is included in main.c,
instead of being linked.
This feature is supported by the usb driver, but needs "USB_PUBLIC" to be defined
as static.
Even when it is not really used in the code,
"static usbMsgLen_t usbFunctionDescriptor(struct usbRequest *rq)" isn't optimized
out at this level of compilation.
Since it is never defined, the warning pops up.

Fix: Define a dummy function which becomes optimized out later.

Signed-off-by: Stephan Baerwolf <stephan.baerwolf@tu-ilmenau.de>