fix wrong free-for-use MAGIC value example
Until now, there was a bit-error in the example for
ATmega8 and ATmega328p:
Accidentally bit "W" was set (meaning active watchdog)
and MSB of "X" was unset (resulting 12MHz instead 16MHz).

Of course the example was for disabled watchdog and 16MHz.
So bit "W" is fixed to unset and MSB of "X" is set.

Signed-off-by: Stephan Baerwolf <stephan@matrixstorm.com>